requirements.txt
aiogram>=3.0.0
vk_api>=11.9.0
google-generativeai>=0.8.0
requests>=2.28.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
